web - extract useEffect login in component/features/matches/matches-list to custom hook
